+++ src/main/java/com/pepero/jcb/api/MoveAnnotation.java
@@ -1,18 +1,28 @@
 package com.pepero.jcb.api;
 
 /**
- * Move annotation for pgn annotation
+ * Move annotation for pgn annotation <br>
+ * If you want to change annotation, use {@link ChessGame#setCurrentMoveCal(String)}, {@link ChessGame#setCurrentMoveClockMilliSeconds(long)}, etc.
  */
 public class MoveAnnotation {
-    public String eval; // engine eval data
-    public String csl; // square light data
-    public String cal; // arrow light data
+    String eval;
+    String csl;
+    String cal;
+    String comment;
+    String nag;
+    String clk;
+    String timeStamp;
 
-    public String comment; // comment data
-    public String nag; // nag data like $2 (?), $3 (!!)
+    public String getEval() { return eval; }
+    public String getCsl() { return csl; }
+    public String getCal() { return cal; }
+    public String getComment() { return comment; }
+    public String getNag() { return nag; }
+    public String getClk() { return clk; }
+    public String getTimeStamp() { return timeStamp; }
 
-    public String clk; // clock data
-    public String timeStamp; // time stamp (time elapsed)
-
-    public boolean isEmpty() { return eval == null && csl == null && cal == null; }
+    public boolean isEmpty() {
+        return eval == null && csl == null && cal == null
+                && comment == null && nag == null && clk == null && timeStamp == null;
+    }
 }